Fix old Fleegix lib code for template adapters
Created new command `scaffold` that generates a complete scaffold

Model, Resource and Scaffold now have a new set of options that are model properties

ex: geddy model user name admin:boolean
The example above generates a user model with the given properties.
Generated models have a default ID property with a 'string' type and if a custom ID property is given
it overrides the default.
If no type is given then it defaults to 'string'

Note: This commit is mainly just to fix the fleegix errors found. Closes #124