gen: add simple passing test for the model
As a beginner I don't know how the model tests work and what
are the best practices. This creates a simple first test so
that every beginner has an example what might be a good idea
in these tests.

As a developer which is used to geddy and testing,
I get most of the boilerplate for my first test of model validations
generated, which makes me happy.