Add a "POST" example for request
It needs to set 'Content-Type' in headers explicitly, when use geddy.request to do a "POST" request.
Or else the data parameters couldn't be passed in. And there is no 'POST' example there.
So add an example for 'POST' request. it's related to Issue #604 (https://github.com/geddy/geddy/issues/604)