Rewrite escapeRegExpChars()
Faster and clearer. I don't know why '/' was in the list of characters
needed to be escaped, so I skipped it.

Similar to Sindre Sorhus' escape-string-regexp, but accepts non-string
argument. Plus the overhead of including a new module is probably greater
than the advantage that comes with using this new implementation anyway.