Fix .eslintignore and lint errors (#207)
The .eslintignore was misconfigured. In an attempt not to validate
the browserified bundles, we added `ejs.js` to the .eslintignore.
This not only ignored the bundle, it also ignored `lib/ejs.js` &
`test/ejs.js`. This line was corrected to read `/ejs.js`.

As a result of this, quite a few linting errors needed fixed.